chronicle/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.10.0
django-filter==21.1
drf-yasg==1.17.1
djangorestframework-simplejwt==4.6.0
django-redis==5.2.0
mysqlclient==2.0.3